# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Code Graph Knowledge Service",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/knowledge/query": {
      "post": {
        "tags": [
          "knowledge-query"
        ],
        "summary": "执行 GraphRAG 查询",
        "description": "对选定知识源发起 GraphRAG 查询，返回答案、置信度与证据锚点信息。",
        "operationId": "query_knowledge_graph_api_v1_knowledge_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphRAGQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "查询成功并返回结构化回答",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphRAGQueryResponse"
                }
              }
            }
          },
          "400": {
            "description": "查询请求无效或底层处理失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphRAGErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "RBAC：权限不足"
          },
          "404": {
            "description": "指定的知识源不存在"
          },
          "408": {
            "description": "查询超时",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphRAGErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphRAGErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/knowledge/search": {
      "post": {
        "tags": [
          "General"
        ],
        "summary": "Search Knowledge",
        "description": "Search similar nodes in knowledge base",
        "operationId": "search_knowledge_api_v1_knowledge_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api__neo4j_routes__SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/documents": {
      "post": {
        "tags": [
          "General"
        ],
        "summary": "Add Document",
        "description": "Add document to knowledge base",
        "operationId": "add_document_api_v1_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/documents/file": {
      "post": {
        "tags": [
          "General"
        ],
        "summary": "Add File",
        "description": "Add file to knowledge base",
        "operationId": "add_file_api_v1_documents_file_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "file_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/directory": {
      "post": {
        "tags": [
          "General"
        ],
        "summary": "Add Directory",
        "description": "Add directory to knowledge base",
        "operationId": "add_directory_api_v1_documents_directory_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectoryProcessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/sql/parse": {
      "post": {
        "tags": [
          "General"
        ],
        "summary": "Parse Sql",
        "description": "Parse SQL statement",
        "operationId": "parse_sql_api_v1_sql_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SQLParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sql/validate": {
      "post": {
        "tags": [
          "General"
        ],
        "summary": "Validate Sql",
        "description": "Validate SQL syntax",
        "operationId": "validate_sql_api_v1_sql_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SQLParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sql/convert": {
      "post": {
        "tags": [
          "General"
        ],
        "summary": "Convert Sql Dialect",
        "description": "Convert SQL between dialects",
        "operationId": "convert_sql_dialect_api_v1_sql_convert_post",
        "parameters": [
          {
            "name": "sql",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sql"
            }
          },
          {
            "name": "from_dialect",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "From Dialect"
            }
          },
          {
            "name": "to_dialect",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "To Dialect"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sql/parse-schema": {
      "post": {
        "tags": [
          "General"
        ],
        "summary": "Parse Sql Schema",
        "description": "Parse SQL schema with smart auto-detection\n\nAutomatically detects:\n- SQL dialect (Oracle, MySQL, PostgreSQL, SQL Server)  \n- Business domain classification\n- Table relationships and statistics",
        "operationId": "parse_sql_schema_api_v1_sql_parse_schema_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SQLSchemaParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DirectoryProcessRequest": {
        "properties": {
          "directory_path": {
            "type": "string",
            "title": "Directory Path"
          },
          "recursive": {
            "type": "boolean",
            "title": "Recursive",
            "default": true
          },
          "file_patterns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Patterns"
          }
        },
        "type": "object",
        "required": [
          "directory_path"
        ],
        "title": "DirectoryProcessRequest"
      },
...
      "DocumentAddRequest": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": "Untitled"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "DocumentAddRequest"
      },
...
      "EvidenceAnchor": {
        "properties": {
          "source_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Id"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "content_snippet": {
            "type": "string",
            "title": "Content Snippet"
          },
          "relevance_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Relevance Score"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number"
          },
          "section_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Title"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "source_name",
          "content_snippet",
          "relevance_score"
        ],
        "title": "EvidenceAnchor",
        "description": "证据锚点模式。",
        "example": {
          "content_snippet": "导入流程会先校验客户编号是否存在于 customer_core 表……",
          "page_number": 12,
          "relevance_score": 0.87,
          "section_title": "数据导入校验",
          "source_id": "550e8400-e29b-41d4-a716-446655440000",
          "source_name": "客户数据库 · customer_orders 表"
        }
      },
...
      "GraphRAGAnswerPayload": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "面向用户的总体回答"
          },
          "related_entities": {
            "items": {
              "$ref": "#/components/schemas/GraphRelatedEntity"
            },
            "type": "array",
            "title": "Related Entities",
            "description": "关联实体列表"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/GraphEvidenceItem"
            },
            "type": "array",
            "title": "Evidence",
            "description": "证据片段列表"
          },
          "next_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Next Actions",
            "description": "建议的后续操作"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "GraphRAGAnswerPayload",
        "description": "GraphRAG 查询回答主体。",
        "example": {
          "evidence": [
            {
              "id": "commit:98c1",
              "snippet": "Improve login error handling by surfacing audit entries.",
              "source_ref": "98c1f3d",
              "source_type": "commit"
            }
          ],
          "next_actions": [
            "查看这些文件最近的提交详情。",
            "联系相关负责人确认变更计划。"
          ],
          "related_entities": [
            {
              "detail": "提交 98c1 修改了异常处理逻辑",
              "importance": "high",
              "name": "backend/services/auth_service.py",
              "type": "file"
            }
          ],
          "summary": "登录流程最近在 2025-10-23 的提交中进行了错误处理优化。"
        }
      },
      "GraphRAGErrorResponse": {
        "properties": {
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "description": "错误代码"
          },
          "error_message": {
            "type": "string",
            "title": "Error Message",
            "description": "错误信息"
          },
          "query_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Id"
          },
          "processing_time_ms": {
            "type": "integer",
            "title": "Processing Time Ms",
            "description": "处理时间（毫秒）"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "error_message",
          "processing_time_ms"
        ],
        "title": "GraphRAGErrorResponse",
        "description": "GraphRAG 错误响应模式。",
        "example": {
          "error_code": "TIMEOUT",
          "error_message": "查询超时，请尝试简化查询或增加超时时间",
          "processing_time_ms": 30050,
          "query_id": "f19bb9b1-30fe-4c3f-ae2e-25a6ca28f0e3"
        }
      },
      "GraphRAGQueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Query",
            "description": "查询文本"
          },
          "source_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ids",
            "description": "限制查询的知识源ID列表"
          },
          "max_results": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Results",
            "description": "最大返回结果数",
            "default": 10
          },
          "include_evidence": {
            "type": "boolean",
            "title": "Include Evidence",
            "description": "是否包含证据信息",
            "default": true
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 300.0,
            "minimum": 5.0,
            "title": "Timeout Seconds",
            "description": "查询超时时间（秒）",
            "default": 30
          },
          "context_query_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Query Id",
            "description": "上一轮查询的 query_id，命中缓存时用于多轮上下文"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "GraphRAGQueryRequest",
        "description": "GraphRAG 查询请求模式。",
        "example": {
          "context_query_id": "ab2f89c3-7f02-4f39-a815-0123456789ab",
          "include_evidence": true,
          "max_results": 5,
          "query": "总结客户资料导入流程涉及的关键表和字段",
          "source_ids": [
            "550e8400-e29b-41d4-a716-446655440000",
            "3d6f0a1c-49f1-4c61-86df-d2dfd52a9bfa"
          ],
          "timeout_seconds": 30
        }
      },
      "GraphRAGQueryResponse": {
        "properties": {
          "answer": {
            "$ref": "#/components/schemas/GraphRAGAnswerPayload",
            "description": "结构化回答"
          },
          "confidence_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Score",
            "description": "答案置信度"
          },
          "evidence_anchors": {
            "items": {
              "$ref": "#/components/schemas/EvidenceAnchor"
            },
            "type": "array",
            "title": "Evidence Anchors",
            "description": "原始证据锚点列表"
          },
          "raw_messages": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Messages",
            "description": "LLM 或检索返回的原始消息"
          },
          "sources_queried": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources Queried",
            "description": "实际查询的知识源标识列表"
          },
          "processing_time_ms": {
            "type": "integer",
            "title": "Processing Time Ms",
            "description": "处理时间（毫秒）"
          },
          "query_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Id",
            "description": "查询ID"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues",
            "description": "上下文构建中的警告信息"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "confidence_score",
          "processing_time_ms"
        ],
        "title": "GraphRAGQueryResponse",
        "description": "GraphRAG 查询响应模式。",
        "example": {
          "answer": {
            "evidence": [
              {
                "id": "commit:f19bb9b1",
                "snippet": "Refactor importer error handling to capture audit entries.",
                "source_ref": "f19bb9b1",
                "source_type": "commit"
              }
            ],
            "next_actions": [
              "查看这些文件最近的提交详情。",
              "联系相关负责人确认变更计划。"
            ],
            "related_entities": [
              {
                "detail": "在提交 f19bb9b1 中更新",
                "importance": "high",
                "name": "backend/services/importer.py",
                "type": "file"
              },
              {
                "detail": "Refactor importer error handling",
                "importance": "high",
                "name": "f19bb9b1",
                "type": "commit"
              }
            ],
            "summary": "导入流程最近的改动集中在 `backend/services/importer.py`，提交 f19bb9b1 调整了错误处理。"
          },
          "confidence_score": 0.78,
          "issues": [],
          "processing_time_ms": 1432,
          "query_id": "f19bb9b1-30fe-4c3f-ae2e-25a6ca28f0e3",
          "sources_queried": [
            "550e8400-e29b-41d4-a716-446655440000"
          ]
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SQLParseRequest": {
        "properties": {
          "sql": {
            "type": "string",
            "title": "Sql"
          },
          "dialect": {
            "type": "string",
            "title": "Dialect",
            "default": "mysql"
          }
        },
        "type": "object",
        "required": [
          "sql"
        ],
        "title": "SQLParseRequest"
      },
      "SQLSchemaParseRequest": {
        "properties": {
          "schema_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Content"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          }
        },
        "type": "object",
        "title": "SQLSchemaParseRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "api__neo4j_routes__SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchRequest"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}